BEL-4546 Fix not passing component args to system under test
Some of our tests were failing because we were reconstructing sut without passing through the component kwargs. Fix this.

Also, take the opportunity to move the shared examples out into their own file.
